feat(desktop): switch to Flat Remix theme
This includes the GTK, icon, and shell themes from Flat Remix. By default we use the Dark variant.
We also use the Dash to Panel Gnome extension as recommended by Flat Remix.

BREAKING CHANGE: if you were used to the default look of Ubuntu, this is a major change.

fix #17